+++ src/alsyntaxmodifiers/lockRemovedFieldsCaptionsModifier.ts
@@ -1,0 +1,10 @@
+import { DevToolsExtensionContext } from "../devToolsExtensionContext";
+import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
+
+export class LockRemovedFieldsCaptionsModifier extends WorkspaceCommandSyntaxModifier {
+
+    constructor(context: DevToolsExtensionContext) {
+        super(context, "Lock Removed Table Fields Captions", "lockRemovedFieldsCaptions");
+    }
+
+}